# docs/04-api/endpoints/openapi-spec.yaml
openapi: 3.1.0
info:
  title: ToolboxAI Roblox Environment
  description: >-
    AI-Powered Educational Roblox Environment - Generate immersive educational
    content with AI agents
  version: 1.0.0
paths:
...
  /api/v1/content/generate:
    post:
      tags:
        - Content API
      summary: Api Generate Content
      description: >-
        API v1 content generation endpoint with proper authorization header
        handling
      operationId: api_generate_content_api_v1_content_generate_post
      parameters:
        - name: authorization
          in: query
          required: false
          schema:
            type: string
            title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /content/{content_id}:
    get:
      tags:
        - Content Management
      summary: Get Content
      description: Retrieve previously generated content by ID
      operationId: get_content_content__content_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            title: Content Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContentRequest:
      properties:
        subject:
          $ref: '#/components/schemas/SubjectType'
          description: Educational subject
        grade_level:
          type: integer
          maximum: 12
          minimum: 1
          title: Grade Level
          description: Grade level (K-12)
        learning_objectives:
          items:
            $ref: '#/components/schemas/LearningObjective'
          type: array
          maxItems: 10
          minItems: 1
          title: Learning Objectives
          description: Learning objectives
        environment_type:
          $ref: '#/components/schemas/EnvironmentType'
          description: Type of Roblox environment
          default: classroom
        terrain_size:
          $ref: '#/components/schemas/TerrainSize'
          description: Size of the terrain
          default: medium
        include_quiz:
          type: boolean
          title: Include Quiz
          description: Whether to include a quiz
          default: true
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
          description: Content difficulty
          default: medium
        duration_minutes:
          type: integer
          maximum: 180
          minimum: 5
          title: Duration Minutes
          description: Expected activity duration in minutes
          default: 30
        max_students:
          type: integer
          maximum: 100
          minimum: 1
          title: Max Students
          description: Maximum number of students
          default: 30
        accessibility_features:
          type: boolean
          title: Accessibility Features
          description: Include accessibility features
          default: false
        multilingual:
          type: boolean
          title: Multilingual
          description: Support multiple languages
          default: false
        custom_requirements:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Custom Requirements
          description: Additional custom requirements
      type: object
      required:
        - subject
        - grade_level
        - learning_objectives
      title: ContentRequest
      description: Request model for generating educational content
    ContentResponse:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the operation was successful
        message:
          type: string
          title: Message
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: Response timestamp
        request_id:
          type: string
          title: Request Id
          description: Unique request identifier
        content:
          additionalProperties: true
          type: object
          title: Content
          description: Generated content data
        scripts:
          items:
            $ref: '#/components/schemas/GeneratedScript'
          type: array
          title: Scripts
          description: Generated Lua scripts
        terrain:
          anyOf:
            - $ref: '#/components/schemas/TerrainConfiguration'
            - type: 'null'
          description: Terrain configuration
        game_mechanics:
          anyOf:
            - $ref: '#/components/schemas/GameMechanics'
            - type: 'null'
          description: Game mechanics settings
        estimated_build_time:
          type: integer
          title: Estimated Build Time
          description: Estimated build time in minutes
          default: 0
        resource_requirements:
          additionalProperties:
            type: integer
          type: object
          title: Resource Requirements
          description: Resource requirements
        content_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Content Id
          description: Unique identifier for the generated content
      type: object
      required:
        - success
        - message
        - content
      title: ContentResponse
      description: Response model for content generation
...
    DifficultyLevel:
      type: string
      enum:
        - easy
        - medium
        - hard
        - expert
      title: DifficultyLevel
      description: Quiz and content difficulty levels
    EnvironmentType:
      type: string
      enum:
        - classroom
        - laboratory
        - outdoor
        - historical
        - futuristic
        - underwater
        - space_station
        - fantasy
        - custom
      title: EnvironmentType
      description: Types of Roblox environments
    GameMechanics:
      properties:
        movement_enabled:
          type: boolean
          title: Movement Enabled
          description: Allow player movement
          default: true
        chat_enabled:
          type: boolean
          title: Chat Enabled
          description: Enable chat system
          default: true
        collision_detection:
          type: boolean
          title: Collision Detection
          description: Enable collision detection
          default: true
        gravity_enabled:
          type: boolean
          title: Gravity Enabled
          description: Enable gravity
          default: true
        respawn_enabled:
          type: boolean
          title: Respawn Enabled
          description: Enable player respawning
          default: true
        team_mode:
          type: boolean
          title: Team Mode
          description: Enable team-based activities
          default: false
      type: object
      title: GameMechanics
      description: Game mechanics configuration
    GeneratedScript:
      properties:
        name:
          type: string
          title: Name
          description: Script name
        content:
          type: string
          title: Content
          description: Lua script content
        script_type:
          type: string
          enum:
            - server
            - client
            - module
          title: Script Type
          description: Type of script
        dependencies:
          items:
            type: string
          type: array
          title: Dependencies
          description: Required dependencies
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
          description: Script description
      type: object
      required:
        - name
        - content
        - script_type
      title: GeneratedScript
      description: Generated Lua script for Roblox
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LearningObjective:
      properties:
        id:
          anyOf:
            - type: string
            - type: 'null'
          title: Id
          description: Unique identifier
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Objective title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Detailed description
        bloom_level:
          anyOf:
            - type: string
            - type: 'null'
          title: Bloom Level
          description: Bloom's taxonomy level
        measurable:
          type: boolean
          title: Measurable
          description: Whether the objective is measurable
          default: true
      type: object
      required:
        - title
      title: LearningObjective
      description: Individual learning objective
...
    SubjectType:
      type: string
      enum:
        - Mathematics
        - Science
        - History
        - English
        - Art
        - Geography
        - Computer Science
        - Physics
        - Chemistry
        - Biology
      title: SubjectType
      description: Supported educational subjects
    TerrainConfiguration:
      properties:
        material:
          type: string
          title: Material
          description: Primary terrain material
          default: Grass
        size:
          $ref: '#/components/schemas/TerrainSize'
          description: Terrain size
          default: medium
        features:
          items:
            type: string
          type: array
          title: Features
          description: Terrain features (hills, water, etc.)
        biome:
          type: string
          title: Biome
          description: Terrain biome
          default: temperate
        elevation_map:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Elevation Map
          description: Custom elevation data
      type: object
      title: TerrainConfiguration
      description: Terrain generation configuration
    TerrainSize:
      type: string
      enum:
        - small
        - medium
        - large
        - xlarge
      title: TerrainSize
      description: Terrain size options
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer